+++ samples/Cpp/TestCppClientSSL/TestCppClientSSl.h
@@ -66,7 +66,7 @@ public:
 	void connectionClosed();
 	void updateAccountValue(const std::string& key, const std::string& val,
 		const std::string& currency, const std::string& accountName);
-	void updatePortfolio(const Contract& contract, int position,
+	void updatePortfolio(const Contract& contract, double position,
 		double marketPrice, double marketValue, double averageCost,
 		double unrealizedPNL, double realizedPNL, const std::string& accountName);
 	void updateAccountTime(const std::string& timeStamp);
@@ -100,7 +100,7 @@ public:
 	void tickSnapshotEnd(int reqId);
 	void marketDataType(TickerId reqId, int marketDataType);
 	void commissionReport( const CommissionReport& commissionReport);
-	void position( const std::string& account, const Contract& contract, int position, double avgCost);
+	void position( const std::string& account, const Contract& contract, double position, double avgCost);
 	void positionEnd();
 	void accountSummary( int reqId, const std::string& account, const std::string& tag, const std::string& value, const std::string& curency);
 	void accountSummaryEnd( int reqId);
